ai-agency-suite/requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
aiosqlite>=0.19.0

# Vector Database
chromadb>=0.4.18
sentence-transformers>=2.2.2

# AI/LLM
ollama>=0.1.0
groq>=0.4.0
openai>=1.3.0

# Web Scraping & Automation
playwright>=1.40.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
httpx>=0.25.0

# Video Processing
moviepy>=1.0.3
Pillow>=10.1.0
opencv-python>=4.8.0

# Text-to-Speech
gTTS>=2.4.0
pyttsx3>=2.90

# Social Media APIs
instagrapi>=2.0.0
google-api-python-client>=2.108.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
facebook-sdk>=3.1.0

# Scheduling
apscheduler>=3.10.0
schedule>=1.2.0

# Dashboard
streamlit>=1.28.0
plotly>=5.18.0
pandas>=2.1.0

# Utilities
python-multipart>=0.0.6
aiofiles>=23.2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
loguru>=0.7.2
rich>=13.7.0
typer>=0.9.0
jinja2>=3.1.2

# Image Generation (Optional)
# stability-sdk>=0.8.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Development
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0
